Update Kinesis using AWS SDK2
