--- interlok-aws-kinesis/src/main/java/com/adaptris/aws/kinesis/KinesisProducerWrapper.java
@@ -1,9 +1,0 @@
-package com.adaptris.aws.kinesis;
-
-import com.amazonaws.services.kinesis.producer.KinesisProducer;
-
-@FunctionalInterface
-public interface KinesisProducerWrapper {
-
-  KinesisProducer kinesisProducer() throws Exception;
-}
