+++ interlok-aws-kinesis/src/main/java/com/adaptris/aws2/kinesis/KinesisProducerWrapper.java
@@ -1,0 +1,9 @@
+package com.adaptris.aws2.kinesis;
+
+import com.amazonaws.services.kinesis.producer.KinesisProducer;
+
+@FunctionalInterface
+public interface KinesisProducerWrapper {
+
+  KinesisProducer kinesisProducer() throws Exception;
+}
