+++ interlok-aws2-kinesis-sdk/README.md
@@ -1,0 +1,3 @@
+# interlok-aws-kinesis-sdk
+
+For localstack testing a docker-compose is provided, which monitors the stream.
