+++ interlok-aws2-kinesis-sdk/build.gradle
@@ -1,0 +1,99 @@
+import org.apache.tools.ant.taskdefs.condition.Os
+import org.apache.tools.ant.filters.*
+
+ext {
+  componentName = 'Interlok AWS2/Kinesis SDK'
+  componentDesc = "Components that interact AWS Kinesis data streams via SDK"
+  delombokTargetDir = new File("${project.projectDir}/src/main/generated")
+  awsSDKVersion = "2.29.9"
+  jacksonVersion = '2.18.1'
+}
+
+dependencies {
+  api (platform("com.fasterxml.jackson:jackson-bom:$jacksonVersion"))
+  api (platform("software.amazon.awssdk:bom:$awsSDKVersion"))
+  api ("software.amazon.awssdk:kinesis")
+  api ("software.amazon.awssdk:sdk-core")
+  // Dependency to platform/BOM
+  api ("com.fasterxml.jackson.core:jackson-databind")
+  api ("com.fasterxml.jackson.core:jackson-core")
+  api ("com.fasterxml.jackson.core:jackson-annotations")
+  api ("com.fasterxml.jackson.dataformat:jackson-dataformat-cbor")
+  api project(':interlok-aws2-common')
+}
+
+
+task javadocJar(type: Jar, dependsOn: javadoc) {
+    classifier = 'javadoc'
+    from javadoc.destinationDir
+}
+
+task examplesJar(type: Jar, dependsOn: test) {
+    classifier = 'examples'
+    from new File(buildDir, '/examples')
+}
+
+task sourcesJar(type: Jar, dependsOn: classes) {
+    classifier = 'sources'
+    from sourceSets.main.extensions.delombokTask
+}
+jar {
+  manifest {
+    attributes("Built-By": System.getProperty('user.name'),
+              "Build-Jdk": System.getProperty('java.version'),
+              "Implementation-Title": componentName,
+              "Implementation-Version": project.version,
+              "Implementation-Vendor-Id": project.group,
+              "Implementation-Vendor": organizationName)
+  }
+}
+
+artifacts {
+    archives javadocJar
+    archives examplesJar
+    archives sourcesJar
+}
+
+publishing {
+  publications {
+      mavenJava(MavenPublication) {
+          from components.java
+
+          artifact javadocJar { classifier "javadoc" }
+          artifact examplesJar { classifier "examples" }
+          artifact sourcesJar { classifier "sources" }
+          pom.withXml {
+            asNode().appendNode("name", componentName)
+            asNode().appendNode("description", "Services that interact with AWS Kinesis Streams")
+	        asNode().appendNode("url", "https://interlok.adaptris.net/interlok-docs/#/")
+            def properties = asNode().appendNode("properties")
+            properties.appendNode("target", "3.12.1+")
+            properties.appendNode("tags", "aws,kinesis")
+            properties.appendNode("license", "false")
+            properties.appendNode("readme", "https://github.com/adaptris/interlok-aws/raw/develop/README.md")
+            properties.appendNode("repository", "https://github.com/adaptris/interlok-aws")
+          }
+      }
+  }
+  repositories {
+    maven {
+      credentials {
+        username repoUsername
+        password repoPassword
+      }
+      url mavenPublishUrl
+    }
+  }
+}
+
+
+delombok {
+  target = delombokTargetDir
+}
+
+task deleteGeneratedFiles(type: Delete) {
+  delete file(testResourcesDir() + "/unit-tests.properties"), file(testResourcesDir() + "/unit-tests.properties.resolved"), delombokTargetDir
+}
+
+clean.dependsOn deleteGeneratedFiles
+processTestResources.dependsOn copyUnitTestProperties
