+++ settings.gradle
@@ -11,3 +11,4 @@ include 'interlok-aws2-s3'
 include 'interlok-aws2-kms'
 include 'interlok-aws2-sqs'
 include 'interlok-aws2-sns'
+include 'interlok-aws2-kinesis-sdk'
